Forgot to commit important changes to 2bl! Added Uncommitted changes feature. Added Menu entry in Info menu to list uncommitted changes.

+++ lib/LPCMod/BootLPCMod.h
@@ -43,7 +43,7 @@ void LPCMod_LCDBankString(char * string, u8 bankID);
 int LPCMod_ReadCFGFromHDD(_LPCmodSettings *LPCmodSettingsPtr, _settingsPtrStruct *settingsStruct);
 int LPCMod_SaveCFGToHDD(void);
 
-int LPCMod_CountNumberOfChangesInSettings(void);
+u8 LPCMod_CountNumberOfChangesInSettings(void);
 
 
 #endif // _BootLPCMod_H_
